# openapi.yaml
openapi: "3.0.3"
info:
  title: nano-dataset-manager
  description: |
    Dataset intake service with disk storage, SQLite metadata,
    zip upload support, and Hugging Face download support using
    user access tokens.
  version: "1.0.0"

paths:
  /datasets/upload:
    post:
      summary: Upload a dataset as a zip file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, name]
              properties:
                file:
                  type: string
                  format: binary
                  description: Zip file containing the dataset files
                name:
                  type: string
                  description: Human-readable dataset name
      responses:
        "201":
          description: Dataset created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "400":
          description: Invalid request (missing fields, invalid form)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Processing error (invalid zip, extraction failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /datasets/from-hf:
    post:
      summary: Download a dataset from Hugging Face
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo_id, token, name]
              properties:
                repo_id:
                  type: string
                  description: Hugging Face repository ID, e.g. "username/dataset-name"
                token:
                  type: string
                  description: Hugging Face user access token (read scope)
                name:
                  type: string
                  description: Human-readable dataset name
      responses:
        "201":
          description: Dataset created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "400":
          description: Invalid request (missing fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Processing error (API error, download failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /datasets:
    get:
      summary: List all datasets
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"

  /datasets/{id}:
    get:
      summary: Get a dataset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dataset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "404":
          description: Dataset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a dataset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Dataset deleted (no content)
        "404":
          description: Dataset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

components:
  schemas:
    Dataset:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique dataset identifier
        name:
          type: string
          description: Human-readable dataset name
        source_type:
          type: string
          enum: [upload, huggingface]
          description: How the dataset was ingested
        source_ref:
          type: string
          description: Original filename (upload) or Hugging Face repo ID
        local_path:
          type: string
          description: Path to the dataset on disk relative to the server
        size_bytes:
          type: integer
          description: Total size of dataset files in bytes
        status:
          type: string
          enum: [pending, ready, failed]
          description: Processing status
        created_at:
          type: string
          format: date-time
          description: When the dataset was created
        updated_at:
          type: string
          format: date-time
          description: When the dataset was last updated
        error_msg:
          type: string
          description: Error message if status is "failed"
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error description